Macos/requirements.txt
albumentations==2.0.6
coremltools==8.3.0
Flask==3.1.0
ipython==8.12.3
matplotlib==3.6.2
numpy==1.23.5
onnx==1.13.0
onnxruntime==1.13.1
onnxsim==0.4.36
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
openvino==2025.1.0
pafy==0.5.5
pandas==1.5.2
Pillow==9.3.0
Pillow==11.2.1
psutil==5.9.4
pycocotools==2.0.8
pyserial==3.5
PyYAML==6.0
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==61.2.0
tensorflow_macos==2.10.0
tensorflowjs==4.22.0
tensorrt==10.9.0.34
tflite_runtime==2.14.0
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.0
websocket_client==1.4.2
